# Cache beatmaps page json response for guest