@notbakaneko 2022-10-05
The potential issue I can see is when the js-side components get updated and they try to render the cached json response.