@nanaya 2022-10-05
I think you mean the json transformer getting updated. But then again it should be relatively stable nowadays and things are updated in backward compatible ways.