@notbakaneko 2022-10-05
No, I do mean js; updated transformer doesn't matter if there's nothing to read the output.